# web-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Trading terminal API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/instruments/{symbol}/quote": {
      "get": {
        "summary": "Live quote with 24h stats, 52-week range, all-time high and market data",
        "tags": [
          "market"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9]+-[A-Z]+$"
            },
            "in": "path",
            "name": "symbol",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asOf": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "live",
                        "replay",
                        "drill",
                        "offline"
                      ]
                    },
                    "source": {
                      "type": "string"
                    },
                    "quote": {
                      "type": "object",
                      "properties": {
                        "symbol": {
                          "type": "string"
                        },
                        "price": {
                          "type": "string"
                        },
                        "change24h": {
                          "nullable": true,
                          "type": "string"
                        },
                        "changePct24h": {
                          "nullable": true,
                          "type": "string"
                        },
                        "open24h": {
                          "nullable": true,
                          "type": "string"
                        },
                        "high24h": {
                          "nullable": true,
                          "type": "string"
                        },
                        "low24h": {
                          "nullable": true,
                          "type": "string"
                        },
                        "previousClose": {
                          "nullable": true,
                          "type": "string"
                        },
                        "volume24h": {
                          "nullable": true,
                          "type": "string"
                        },
                        "marketCap": {
                          "nullable": true,
                          "type": "string"
                        },
                        "rank": {
                          "nullable": true,
                          "type": "number"
                        },
                        "high52w": {
                          "nullable": true,
                          "type": "string"
                        },
                        "low52w": {
                          "nullable": true,
                          "type": "string"
                        },
                        "allTimeHigh": {
                          "nullable": true,
                          "type": "string"
                        },
                        "circulatingSupply": {
                          "nullable": true,
                          "type": "string"
                        },
                        "maxSupply": {
                          "nullable": true,
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "sources": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "symbol",
                        "price",
                        "change24h",
                        "changePct24h",
                        "open24h",
                        "high24h",
                        "low24h",
                        "previousClose",
                        "volume24h",
                        "marketCap",
                        "rank",
                        "high52w",
                        "low52w",
                        "allTimeHigh",
                        "circulatingSupply",
                        "maxSupply",
                        "updatedAt",
                        "sources"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "asOf",
                    "mode",
                    "source",
                    "quote"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/instruments/{symbol}/candles": {
      "get": {
        "summary": "OHLCV candles. Use `range` (chart tabs) or `from`/`to`; interval is chosen for you if omitted.",
        "tags": [
          "market"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "1m",
                "5m",
                "15m",
                "1h",
                "4h",
                "1d"
              ]
            },
            "in": "query",
            "name": "interval",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "1D",
                "1W",
                "1M",
                "3M",
                "1Y",
                "5Y",
                "All"
              ]
            },
            "in": "query",
            "name": "range",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
            },
            "in": "query",
            "name": "to",
            "required": false
          },
          {
            "schema": {
              "default": 500,
              "type": "integer",
              "minimum": 1,
              "maximum": 2000
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Z0-9]+-[A-Z]+$"
            },
            "in": "path",
            "name": "symbol",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "asOf": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "live",
                        "replay",
                        "drill",
                        "offline"
                      ]
                    },
                    "source": {
                      "type": "string"
                    },
                    "symbol": {
                      "type": "string"
                    },
                    "interval": {
                      "type": "string"
                    },
                    "candles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "t": {
                            "type": "string"
                          },
                          "o": {
                            "type": "string"
                          },
                          "h": {
                            "type": "string"
                          },
                          "l": {
                            "type": "string"
                          },
                          "c": {
                            "type": "string"
                          },
                          "v": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "t",
                          "o",
                          "h",
                          "l",
                          "c",
                          "v",
                          "source"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "asOf",
                    "mode",
                    "source",
                    "symbol",
                    "interval",
                    "candles"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}